docs: Refactor Appearance demo and docs
Move and rename the HeroUI appearance demo into the theme demo area, delete the old appearance MDX doc, and update references. Replace component preview name to heroui-theme-appearance, update demos index mapping and keys, and adjust demo imports to use fumadocs-ui/provider/base for useTheme and the shared authClient. Also reorder plugin pages arrays for heroui and shadcn meta.json files to standardize page ordering.
