+++ apps/docs/content/docs/heroui/components/meta.json
@@ -20,7 +20,6 @@
     "settings/account/account-settings",
     "settings/account/user-profile",
     "settings/account/change-email",
-    "settings/account/appearance",
     "settings/security/security-settings",
     "settings/security/change-password",
     "settings/security/linked-accounts",
