--- apps/docs/content/docs/heroui/components/settings/account/appearance.mdx
@@ -1,51 +1,0 @@
----
-title: <Appearance />
-description: A theme selector card with visual theme previews for system, light, and dark themes.
----
-
-## Usage
-
-<ComponentPreview name="heroui-appearance" />
-
-```tsx file=<rootDir>/src/demos/heroui/settings/account/appearance.tsx
-```
-
-## Theme Configuration
-
-The Appearance section reads `appearance` from `useAuth()`. It only renders when `setTheme` is defined and `themes` has at least one entry. Defaults from the library supply `themes: ["system", "light", "dark"]` via config merge, so you typically pass **`theme` and `setTheme`** from your theme library (for example [next-themes](https://github.com/pacocoursey/next-themes)).
-
-Wrap the app with `ThemeProvider` from `next-themes` (for example in your root layout), then wire the same values into `AuthProvider`:
-
-```tsx
-import { AuthProvider } from "@better-auth-ui/heroui"
-import { useRouter } from "next/navigation"
-import { useTheme } from "next-themes"
-
-import { authClient } from "@/lib/auth-client"
-
-function Providers({ children }: { children: React.ReactNode }) {
-  const router = useRouter()
-  const { theme, setTheme } = useTheme()
-
-  return (
-    <AuthProvider
-      authClient={authClient}
-      navigate={({ to, replace }) =>
-        replace ? router.replace(to) : router.push(to)
-      }
-      appearance={{
-        theme,
-        setTheme
-        // Optional: override which options appear (default: system, light, dark)
-        // themes: ["light", "dark"],
-      }}
-    >
-      {children}
-    </AuthProvider>
-  )
-}
-```
-
-## Props
-
-<auto-type-table path="../../../../../../../../packages/heroui/src/components/theme/appearance.tsx" name="AppearanceProps" />
