+++ apps/docs/content/docs/heroui/plugins/meta.json
@@ -1,5 +1,5 @@
 {
   "icon": "Plug",
   "defaultOpen": true,
-  "pages": ["magic-link", "passkey", "multi-session", "theme"]
+  "pages": ["magic-link", "multi-session", "passkey", "theme"]
 }
