+++ apps/docs/content/docs/heroui/plugins/theme.mdx
@@ -79,7 +79,7 @@ function Providers({ children }: { children: ReactNode }) {
 
 ### `<Appearance />`
 
-<ComponentPreview name="heroui-appearance" />
+<ComponentPreview name="heroui-theme-appearance" />
 
 The `<Appearance />` card is automatically rendered in `<AccountSettings />` when the plugin is registered.
