+++ apps/docs/content/docs/shadcn/plugins/meta.json
@@ -1,5 +1,5 @@
 {
   "icon": "Plug",
   "defaultOpen": true,
-  "pages": ["magic-link", "passkey", "multi-session"]
+  "pages": ["magic-link", "multi-session", "passkey"]
 }
