+++ apps/docs/src/demos/heroui/theme/appearance.tsx
@@ -1,11 +1,13 @@
 import { AuthProvider } from "@better-auth-ui/heroui"
 import { Appearance, themePlugin } from "@better-auth-ui/heroui/plugins"
-import { useTheme } from "@heroui/react"
+import { useTheme } from "fumadocs-ui/provider/base"
+
+import { authClient } from "@/lib/auth-client"
 
 export function AppearanceDemo() {
   return (
     <AuthProvider
-      authClient={{} as any}
+      authClient={authClient}
       navigate={() => {}}
       plugins={[themePlugin(useTheme())]}
     >
